pyproject.toml
[tool.poetry]
name = "dteamtest"
version = "0.1.0"
description = ""
authors = ["Nik-BZ-22 <bakhamyt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
python-dotenv = "^1.1.0"
weasyprint = "^65.1"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
